requirements.txt
asgiref==3.5.2
Django==4.1.2
django-cors-headers==3.13.0
django-extensions==3.2.1
djangorestframework==3.14.0
pytz==2022.5
sqlparse==0.4.3
tzdata==2022.5
